Make the properties `long`s like the mastodon schema
The mastodon schema defines these items as bigints in PostgreSQL and while it is unlikely that anyone actually has billions of statuses, followers, etc., at least one account claims to have 97B followers: https://mastodon.adtension.com/@admin

+++ Mastonet/Entities/Account.cs
@@ -110,19 +110,19 @@ public class Account
     /// How many statuses are attached to this account.
     /// </summary>
     [JsonProperty("statuses_count")]
-    public int StatusesCount { get; set; }
+    public long StatusesCount { get; set; }
 
     /// <summary>
     /// The reported followers of this profile.
     /// </summary>
     [JsonProperty("followers_count")]
-    public int FollowersCount { get; set; }
+    public long FollowersCount { get; set; }
 
     /// <summary>
     /// The reported follows of this profile.
     /// </summary>
     [JsonProperty("following_count")]
-    public int FollowingCount { get; set; }
+    public long FollowingCount { get; set; }
 
     // Optional attributes
